- Adds a super-lightweight test framework, "tape", to the project - Adds package.json for node/npm management of the single, dev-only   dependency. - Adds a test directory and a single test file under there targeting one   of the validation functions. To run tests: nodejs ./test/*.js (note:   won't work in Windows cmd shell since dos can't do *nix glob -- need   bash for windows or something). - To make it easier (possible?) to do the unit testing, separates the code   validation functions out into their own file. Both 'code-validation.js'   and 'indeni.js' depend on 'getScriptSections', so I had to pull that out   into a separate file ('common.js') to avoid circular dependency.   'indeni.html' imports the .js files in a careful order to make sure   everything's where it should be for each import. Also, adds a node   "export" statement to 'code-validation.js' so that we can access the   validation functions in the tape test modules via requires(). - Adds tests for spacing around comparison operators and update regex   accordingly: now handles all the AWK comparison operators:  =!<>~ - Rename comparison operator spaces function to make it more clear what   it's doing, and rename test file to agree. - add test-runner.sh to make it easy to run tests from IDE
