+++ code-quality/indeni.html
@@ -2,6 +2,8 @@
 	<head>
 		<script src="https://code.jquery.com/jquery-latest.js"></script>
 		<script src="./tippy.js"></script>
+        <script src="./common.js"></script>
+        <script src="./code-validation.js"></script>
 		<script src="./indeni.js"></script>
         <link href="./indeni.css" rel="stylesheet" type="text/css">
 	</head>
