+++ test/test-runner.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+# bash for Windows call node "nodejs"
+nodejs ./*.js
